Fix warnings

+++ BeagleBone_black/src/sensors/fake_imu.cpp
@@ -129,7 +129,7 @@ bool FakeImu::accCheckTime()
     high_resolution_clock::time_point now = high_resolution_clock::now();
     microseconds time_span = duration_cast<microseconds>(now - imu_ref_time);
 
-    if (time_span.count() < kAccTimeInterval*acc_count) {
+    if (unsigned(time_span.count()) < kAccTimeInterval*acc_count) {
         return false;
     }
 
@@ -142,7 +142,7 @@ bool FakeImu::gyrCheckTime()
     high_resolution_clock::time_point now = high_resolution_clock::now();
     microseconds time_span = duration_cast<microseconds>(now - imu_ref_time);
 
-    if (time_span.count() < kGyrTimeInterval*gyr_count) {
+    if (unsigned(time_span.count()) < kGyrTimeInterval*gyr_count) {
         return false;
     }
